Replace Appium UI tests with FlaUI
